separated server components into separate files
