+++ .eslintrc.js
@@ -1,10 +1,8 @@
-/**
- * These rules enforce the Hack Reactor Style Guide
- *
- * Visit this repo for more information:
- *   https://github.com/reactorcore/eslint-config-hackreactor
- */
-
 module.exports = {
-  extends: './node_modules/eslint-config-hackreactor/index.js'
+ "extends": "airbnb",
+ "env": {
+   "jest": true,
+   "browser": true,
+   "node": true
+ },
 };
